+++ XF/Api/Templater.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TickTackk\DeveloperTools\XF\Api;
+
+use TickTackk\DeveloperTools\XF\Template\TemplaterTrait;
+
+class Templater extends XFCP_Templater
+{
+    use TemplaterTrait;
+}