+++ XF/Mail/Templater.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace TickTackk\DeveloperTools\XF\Mail;
+
+use TickTackk\DeveloperTools\XF\Template\TemplaterTrait;
+
+/**
+ * Class Templater
+ *
+ * Extends \XF\Mail\Templater
+ *
+ * @package TickTackk\DeveloperTools\XF\Mail
+ */
+class Templater extends XFCP_Templater
+{
+    use TemplaterTrait;
+}