+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - **New:** Add `tck-devtools:generate-schema-addon` command to generate schema for every entity at once (#128)
 - **New:** Generate code event listener code using code generator (#130)
 - **Fix:** Permission interface groups are not sorted correctly when building readme (#125)
+- **Fix:** Error: Call to undefined method `XF\Api\Templater::logPermissionError()` (#127)
 
 ## 1.4.3 (`1040370`)